:recycle: refactoring: Update deletarRedeSocial method to throw RuntimeException for not found Rede Social instead of using ResponseStatusException

+++ src/main/java/com/conectados/conectados/services/RedeSocialService.java
@@ -36,6 +36,10 @@ public class RedeSocialService {
     }
 
     public void deletarRedeSocial(Long id) {
+        Optional<RedeSocial> redeSocial = buscarRedeSocialPorId(id);
+        if (redeSocial.isEmpty()) {
+            throw new RuntimeException("Rede Social não encontrada");
+        }
         redeSocialRepository.deleteById(id);
     }
